package refactoring to prepare code to have more than 1 controller type. These are related to REST endpoints.
