+++ src/main/java/br/com/well/rest/controller/rest/ClientApi.java
@@ -1,4 +1,4 @@
-package br.com.well.rest.controller;
+package br.com.well.rest.controller.rest;
 import br.com.well.rest.service.model.ClientModel;
 import io.swagger.annotations.*;
 import org.springframework.http.ResponseEntity;
